requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.6.2
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
PyJWT==2.6.0
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
